+++ application/models/Accessories.php
@@ -48,4 +48,5 @@ class Accessories extends CI_Model {
         return $this->data;
     }
 
+    
 }