+++ application/views/modify.php
@@ -1,10 +1,12 @@
 <tr>
     <!-- MODIFY the next line -->
-    <td>{EquipmentID}</td>
-    <td><input type="text" value="{Category}"/></td>
-    <td><input type="text" value="{Name}"/></td>
-    <td><input type="number" value="{Mobility}"/></td>
-    <td><input type="number" value="{Range}"/></td>
-    <td><input type="number" value="{Power}"/></td>
-    <td><input type="number" value="{Protection}"/></td>
+    <td><input id="id" type="number" value="{EquipmentID}"/></td>
+    <td><input id="category" type="text" value="{Category}"/></td>
+    <td><input id="name" type="text" value="{Name}"/></td>
+    <td><input id="mobility" type="number" value="{Mobility}"/></td>
+    <td><input id="range" type="number" value="{Range}"/></td>
+    <td><input id="power" type="number" value="{Power}"/></td>
+    <td><input id="protection" type="number" value="{Protection}"/></td>
 </tr>
+<div id="submit_btn" onclick="modify_item()" class="btn">Submit</div>
+<script src="/assets/js/modify.js"></script>